+++ client/src/components/Rating/Rating.jsx
@@ -1,0 +1,7 @@
+import {getOneRating} from "../../services/ratings"
+function Rating() {
+  const { id } = useParams();
+  const [isLoaded, setLoaded] = useState(false);
+  return (<div></div>)
+}
+export default Rating
