+++ client/src/containers/MainContainer.jsx
@@ -1,11 +1,20 @@
 import { Switch, Route } from "react-router-dom";
 import Books from "../screens/Books/Books";
 import BookDetail from "../screens/BookDetail/BookDetail";
-import BookCreate from "../screens/BookCreate";
+import BookCreate from "../screens/BookCreate/BookCreate";
 import BookEdit from "../screens/BookEdit/BookEdit";
+import { getAllRatings } from "../services/ratings";
+import {useState, useEffect } from "react"
 
 function MainContainer(props) {
-
+  const [ratings, setRatings] = useState(null);
+  useEffect(() => {
+    const fetchRatings = async () => {
+      const ratingsData = await getAllRatings();
+      setRatings(ratingsData);
+    };
+    fetchRatings();
+  }, []);
   
   return (
     <div>
@@ -17,7 +26,7 @@ function MainContainer(props) {
           <BookCreate />
           </Route>
         <Route path="/books/:id">
-          <BookDetail currentUser={props.currentUser}/>
+          <BookDetail ratings={ratings} currentUser={props.currentUser}/>
         </Route>
         <Route path="/">
          <Books/>
