+++ client/src/screens/BookEdit/BookEdit.css
@@ -55,12 +55,11 @@ Layout {
 
 .save-button {
   font-size: 22px;
-  font-family: Arial, Helvetica, sans-serif;
   height: 40px;
   width: 200px;
   margin-top: 40px;
   margin-left: 300px;
-  background-color: black;
+  color: #845ec2;
   border: none;
   box-shadow: 2px 2px 10px 2px #cecece;
   cursor: pointer;
@@ -125,8 +124,7 @@ Layout {
   height: 40px;
   width: 100px;
   margin-top: 40px;
-  background-color: rgb(8, 143, 143);
-  color: white;
+  background-color: #fefedf;
   border: none;
   box-shadow: 2px 2px 10px 2px #cecece;
   cursor: pointer;
@@ -164,6 +162,7 @@ Layout {
     text-align: center;
     margin-top: 8px;
     font-size: 20px;
+    color: #845ec2;
     font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto",
       "Oxygen", "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans",
       "Helvetica Neue", sans-serif;
@@ -174,8 +173,8 @@ Layout {
     height: 40px;
     width: 100px;
     margin-top: 40px;
-    background-color: rgb(8, 143, 143);
-    color: white;
+    background-color: #fefedf;
+  
     border: none;
     box-shadow: 2px 2px 10px 2px #cecece;
     cursor: pointer;
